# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/admin/tokens": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List all tokens / 列出所有 Token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-array_api_AdminTokenItem"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/tokens/{jti}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Revoke token / 撤销 Token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token jti / token 标识",
                        "name": "jti",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/users": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Root user lists all users.\nroot 用户查看所有用户。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List users / 列出用户",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-array_api_AdminUserItem"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Root creates a normal user.\nroot 创建普通用户。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create user / 创建用户",
                "parameters": [
                    {
                        "description": "request / 请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AdminCreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-api_AdminUserItem"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/users/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Root deletes a user, but root itself is protected.\nroot 删除用户，但 root 自身受保护不可删除。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete user / 删除用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id / 用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/users/{id}/password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Reset user password / 重置用户密码",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id / 用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request / 请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AdminResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.AdminCreateUserRequest": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.AdminResetPasswordRequest": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                }
            }
        },
        "api.AdminTokenItem": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "idle_timeout_seconds": {
                    "type": "integer"
                },
                "issued_at": {
                    "type": "string"
                },
                "jti": {
                    "type": "string"
                },
                "last_seen_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "revoked_at": {
                    "type": "string"
                },
                "revoked_by": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.TokenType"
                },
                "user_id": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "api.AdminUserItem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_root": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "models.TokenType": {
            "type": "string",
            "enum": [
                "web",
                "api"
            ],
            "x-enum-varnames": [
                "TokenTypeWeb",
                "TokenTypeAPI"
            ]
        },
        "response.Envelope-any": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {},
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
        "response.Envelope-api_AdminUserItem": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {
                    "$ref": "#/definitions/api.AdminUserItem"
                },
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
...
        "response.Envelope-array_api_AdminTokenItem": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.AdminTokenItem"
                    }
                },
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
        "response.Envelope-array_api_AdminUserItem": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/response.ErrorCode"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/api.AdminUserItem"
                    }
                },
                "message": {
                    "description": "English message / 英文信息",
                    "type": "string"
                }
            }
        },
...
        "response.ErrorCode": {
            "type": "integer",
            "enum": [
                0,
                10000,
                10001,
                10002,
                10003,
                10004,
                10005
            ],
            "x-enum-varnames": [
                "CodeOK",
                "CodeBadRequest",
                "CodeUnauthorized",
                "CodeForbidden",
                "CodeNotFound",
                "CodeConflict",
                "CodeInternal"
            ]
        }
    }
}